+++ tests/Unit/Web2smsChannelTest.php
@@ -5,7 +5,7 @@ namespace ITPalert\Web2smsChannel\Tests\Unit;
 use Mockery;
 use PHPUnit\Framework\TestCase;
 use Illuminate\Notifications\Notification;
-use ITPalert\Web2sms\Client;
+use ITPalert\Web2sms\Contracts\Client;
 use ITPalert\Web2sms\SMS;
 use ITPalert\Web2smsChannel\Channels\Web2smsChannel;
 use ITPalert\Web2smsChannel\Messages\Web2smsMessage;